v4-client-py-v2/pyproject.toml
[tool.poetry]
name = "dydx-v4-client"
version = "1.1.4"
description = ""
authors = [
    "Saul Martin <saulmartin1@pm.me>",
    "Piotr Piwoński <piwonskp@gmail.com>",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
v4-proto = "^7.0.0"
httpx = "^0.27.0"
websocket-client = "^1.7.0"
bip-utils = "^2.9.3"
typing-extensions = "^4.12.2"
bech32 = "^1.2.0"
pycryptodome = "^3.20.0"
grpcio = "^1.68.0"
coincurve = "^20.0.0"
python-dotenv = "^1.1.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
python-dotenv = "^1.0.1"
twine = "^4.0.2"
black = "^24.4.0"

[tool.poetry.group.examples.dependencies]
nest-asyncio = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
